requirements.txt
click==7.1.2
Flask==1.1.2
future==0.18.2
gevent==20.6.2
greenlet==0.4.16
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
numpy==1.19.0
Pillow==7.1.2
torch==1.5.1
torchvision==0.6.1
Werkzeug==1.0.1
zope.event==4.4
zope.interface==5.1.0
